+++ src/Kdyby/BootstrapFormRenderer/Latte/FormMacros.php
@@ -7,16 +7,17 @@
  *
  * For the full copyright and license information, please view the file license.txt that was distributed with this source code.
  */
-
 namespace Kdyby\BootstrapFormRenderer\Latte;
-use Kdyby;
+
 use Kdyby\BootstrapFormRenderer\BootstrapRenderer;
+use Latte\Compiler;
+use Latte\CompileException;
+use Latte\MacroNode;
+use Latte\Macros\MacroSet;
+use Latte\PhpWriter;
 use Nette;
+use Nette\Bridges\FormsLatte\FormMacros as NetteFormMacros;
 use Nette\Forms\Form;
-use Nette\Latte;
-use Nette\Latte\CompileException;
-use Nette\Latte\MacroNode;
-use Nette\Latte\PhpWriter;
 
 
 /**
@@ -30,7 +31,7 @@ use Nette\Latte\PhpWriter;
  * {/form} as {$form->render('end')}
  * </code>
  *
- * Self-closing form (Latte 2.1 semantics):
+ * Self-closing form (Latte 2.2 semantics):
  *
  * <code>
  * {form name /} as {form name}{/form} (begin + hidden fields + end; no body)
@@ -50,19 +51,19 @@ use Nette\Latte\PhpWriter;
  * </code>
  *
  * Related:
- * - {@see \Nette\Latte\Macros\FormMacros} (Latte 2.1 core form macros)
+ * - {@see \Nette\Bridges\FormsLatte\FormMacros} (Nette forms macros for Latte 2.2)
  * - {@see \Kdyby\BootstrapFormRenderer\BootstrapRenderer} (Bootstrap rendering implementation)
  *
  * @author Filip Procházka <filip@prochazka.su>
  */
-class FormMacros extends Latte\Macros\MacroSet
+class FormMacros extends MacroSet
 {
 
 	/**
-	 * @param \Nette\Latte\Compiler $compiler
-	 * @return \Nette\Latte\Macros\MacroSet|void
+	 * @param \Latte\Compiler $compiler
+	 * @return static
 	 */
-	public static function install(Latte\Compiler $compiler)
+	public static function install(Compiler $compiler)
 	{
 		$me = new static($compiler);
 		$me->addMacro('form', array($me, 'macroFormBegin'), array($me, 'macroFormEnd'));
@@ -78,10 +79,10 @@ class FormMacros extends Latte\Macros\MacroSet
 
 
 	/**
-	 * @param \Nette\Latte\MacroNode $node
-	 * @param \Nette\Latte\PhpWriter $writer
+	 * @param \Latte\MacroNode $node
+	 * @param \Latte\PhpWriter $writer
 	 * @return string
-	 * @throws \Nette\Latte\CompileException
+	 * @throws \Latte\CompileException
 	 */
 	public function macroFormBegin(MacroNode $node, PhpWriter $writer)
 	{
@@ -101,20 +102,20 @@ class FormMacros extends Latte\Macros\MacroSet
 
 
 	/**
-	 * @param \Nette\Latte\MacroNode $node
-	 * @param \Nette\Latte\PhpWriter $writer
+	 * @param \Latte\MacroNode $node
+	 * @param \Latte\PhpWriter $writer
 	 */
 	public function macroFormEnd(MacroNode $node, PhpWriter $writer)
 	{
-		return $writer->write('Nette\Latte\Macros\FormMacros::renderFormEnd($_form)');
+		return $writer->write('Nette\Bridges\FormsLatte\FormMacros::renderFormEnd($_form)');
 	}
 
 
 
 	/**
-	 * @param \Nette\Latte\MacroNode $node
-	 * @param \Nette\Latte\PhpWriter $writer
-	 * @throws \Nette\Latte\CompileException
+	 * @param \Latte\MacroNode $node
+	 * @param \Latte\PhpWriter $writer
+	 * @throws \Latte\CompileException
 	 */
 	public function macroPair(MacroNode $node, PhpWriter $writer)
 	{
@@ -129,9 +130,9 @@ class FormMacros extends Latte\Macros\MacroSet
 
 
 	/**
-	 * @param \Nette\Latte\MacroNode $node
-	 * @param \Nette\Latte\PhpWriter $writer
-	 * @throws \Nette\Latte\CompileException
+	 * @param \Latte\MacroNode $node
+	 * @param \Latte\PhpWriter $writer
+	 * @throws \Latte\CompileException
 	 */
 	public function macroGroup(MacroNode $node, PhpWriter $writer)
 	{
@@ -146,9 +147,9 @@ class FormMacros extends Latte\Macros\MacroSet
 
 
 	/**
-	 * @param \Nette\Latte\MacroNode $node
-	 * @param \Nette\Latte\PhpWriter $writer
-	 * @throws \Nette\Latte\CompileException
+	 * @param \Latte\MacroNode $node
+	 * @param \Latte\PhpWriter $writer
+	 * @throws \Latte\CompileException
 	 */
 	public function macroContainer(MacroNode $node, PhpWriter $writer)
 	{
@@ -201,7 +202,7 @@ class FormMacros extends Latte\Macros\MacroSet
 			$form->render('begin', $args);
 
 		} else {
-			Nette\Latte\Macros\FormMacros::renderFormBegin($form, $args);
+			NetteFormMacros::renderFormBegin($form, $args);
 		}
 	}